+++ modules/utxo-core/package.json
@@ -54,7 +54,11 @@
     "@bitgo/unspents": "^0.48.1",
     "@bitgo/utxo-lib": "^11.5.0",
     "@bitgo/wasm-miniscript": "2.0.0-beta.7",
-    "bip174": "npm:@bitgo-forks/bip174@3.1.0-master.4"
+    "bip174": "npm:@bitgo-forks/bip174@3.1.0-master.4",
+    "lodash": "^4.17.15"
+  },
+  "devDependencies": {
+    "@types/lodash": "^4.14.151"
   },
   "gitHead": "18e460ddf02de2dbf13c2aa243478188fb539f0c"
 }
