+++ modules/utxo-core/src/index.ts
@@ -2,6 +2,7 @@ export * as bip65 from './bip65';
 export * as descriptor from './descriptor';
 export * as testutil from './testutil';
 export * as paygo from './paygo';
+export * from './bip32utils';
 export * from './dustThreshold';
 export * from './Output';
 export * from './xOnlyPubkey';
